refactor: 폴더 구조 수정, 경로 절대경로에 맞게 수정
