+++ src/App.jsx
@@ -1,5 +1,5 @@
 import { AppRouter } from './routes/AppRouter.jsx';
-import { CommonToaster } from './components/toasterConfig.jsx';
+import { CommonToaster } from './components/commonToast/toasterConfig.jsx';
 
 function App() {
   return (
