+++ src/hooks/todayFocus/useTimerAction.hook.js
@@ -1,5 +1,5 @@
 import { useState, useEffect, useCallback } from 'react';
-import { INITIAL_SECONDS } from '../../constants/time.js';
+import { INITIAL_SECONDS } from '@/constants/time.js';
 
 export const useTimerAction = () => {
   const [seconds, setSeconds] = useState(INITIAL_SECONDS);
