+++ src/pages/CreateStudy/CreateStudy.module.css
@@ -1,6 +1,6 @@
 .wrapper {
-  width: 100%;
   min-height: 100%;
   display: flex;
   justify-content: center;
+  background-color: #fff;
 }
