+++ src/routes/AppRouter.jsx
@@ -3,10 +3,10 @@ import { RootLayout } from '@/pages/RootLayout/RootLayout.jsx';
 
 import { Home } from '@/pages/Home/home';
 import { UiPreview } from '@/pages/UiPreview';
-import { TodayFocus } from '@/pages/TodayFocus.jsx';
+import { TodayFocus } from '@/pages/TodayFocus/TodayFocus';
 import { StudyDetail } from '@/pages/StudyDetail/StudyDetail';
 import { TodayHabitPage } from '@/pages/TodayHabitPage/TodayHabitPage';
-// import { CreateStudyPage } from '@/pages/CreateStudy.jsx';
+import { CreateStudyPage } from '@/pages/CreateStudy/CreateStudy';
 
 export const AppRouter = () => {
   return (
@@ -14,14 +14,14 @@ export const AppRouter = () => {
       {/* ✅ 전역 레이아웃 + 공통 Header가 필요한 페이지들 */}
       <Route element={<RootLayout />}>
         <Route path="/" element={<Home />} />
-        {/* <Route path="/create" element={<CreateStudyPage />} /> */}
+        <Route path="/create" element={<CreateStudyPage />} />
         <Route path="/studies/:studyId" element={<StudyDetail />} />
         <Route path="/studies/:studyId/focus" element={<TodayFocus />} />
         <Route path="/studies/:studyId/habits" element={<TodayHabitPage />} />
       </Route>
 
       {/* UI 테스트용 */}
-      <Route path="/priview" element={<UiPreview />} />
+      <Route path="/preview" element={<UiPreview />} />
     </Routes>
   );
 };
